LPS-94775 Rename other webpack files for consistency
In the previous commit we made sure the main config file used the
standard name in all three places that we use wepback.

For the auxiliary config files (used for watch mode and the test
environment, for example), we had inconsistent naming. This commit makes
every config file follow the pattern of "webpack.config.$ENV.js", where
`$ENV` is the environment corresponding to that file, or "common" for
configuration used by more than one environment.
