+++ modules/apps/frontend-js/frontend-js-web/webpack.config.common.js
