+++ modules/apps/frontend-js/frontend-js-web/webpack.config.js
@@ -1,7 +1,7 @@
 const merge = require('webpack-merge');
 const webpack = require('webpack');
 
-const common = require('./webpack.common.js');
+const common = require('./webpack.config.common');
 
 module.exports = merge(
 	common.config,
