+++ modules/apps/segments/segments-web/webpack.config.dev.js
